Pull functions out of TableUtils
These functions do not need to exist as a static method. Put them in the readrows resumption module.
